move response handler test to make HttpConnectionTestCase, add a unit test for the deprecated HttpConnection methods
